+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.2-pre
 * Add usbhid driver
+* Add device generation option
 
 ### 0.1.1
 * Move optional platform override to constructor